+++ pkg/frontend/util.go
@@ -1069,8 +1069,8 @@ type UserInput struct {
 	isRestore     bool
 	// operator account, the account executes restoration
 	// e.g. sys takes a snapshot sn1 for acc1, then restores acc1 from snapshot sn1. In this scenario, sys is the operator account
-	opAccount uint32
-	toAccount uint32
+	opAccount int64
+	toAccount int64
 }
 
 func (ui *UserInput) getSql() string {
